when setting redundancy add size
This allows us to set the approximate size we want to be able to
recover.

for example:
We have a set of files of 1,1MB we can choose to be able to have 100kb
missing.

par2 c -rk100 recovery *

would bring us approximatly that result, it will never be exact since
the size is converted to the percentage recovery needed to achieve the
goal.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>